fixed resolvers (using floats instead of doubles to make sure number are rounded correctly) added tests fixed AST matcher to parse number correctly to avoid comparison failures due to different representations of the same number
this fixes bug 721 (Bug in number token resolver(s))
the JACKS test is now down to 3 failures which are caused by the wrong reprints of Javadoc comments within empty classes

+++ org.emftext.language.java.resource.java/src/org/emftext/language/java/resource/java/analysis/JavaDECIMAL_INTEGER_LITERALTokenResolver.java
@@ -8,7 +8,7 @@ public class JavaDECIMAL_INTEGER_LITERALTokenResolver extends org.emftext.runtim
 	@Override
 	public java.lang.String deResolve(java.lang.Object value, org.eclipse.emf.ecore.EStructuralFeature feature, org.eclipse.emf.ecore.EObject container) {
 		assert container == null || container instanceof DecimalIntegerLiteral;
-		assert value instanceof Integer;
+		assert value instanceof Long;
 		return value.toString();
 	}
 
@@ -17,9 +17,11 @@ public class JavaDECIMAL_INTEGER_LITERALTokenResolver extends org.emftext.runtim
 		assert feature == null || feature.getEContainingClass().equals(LiteralsPackage.eINSTANCE.getIntegerLiteral());
 		
 		JavaDECIMAL_LONG_LITERALTokenResolver.parseToLong(lexem, 10, result);
+		/*
 		Long longValue = (Long) result.getResolvedToken();
 		if (longValue != null) {
 			result.setResolvedToken(new Integer(longValue.intValue()));
 		}
+		*/
 	}
 }
